--- domain/entities/Balance.ts
@@ -1,4 +1,0 @@
-export interface Balance {
-  user: string;
-  balance: number;
-}
